yt-ml-pipeline/requirements.txt
torch
torchvision
transformers
pillow
opencv-python
pytesseract
openai-whisper
clip-by-openai
